when a syscall fails print also its arguments on stderr
this is especially useful to understand why execvp() failed because now
we can "see" the exact command it tried to execute. for example, pgrep
could return more than one pid for a given pattern, after will "wait" on
the first one and then execvp() tries to "run" the second pid failing
badly. e.g.:

    after $(pgrep rsync) systemctl suspend

might actually be:

    after 77789 77790 77791 systemctl suspend

and the error you got before was:

    after: execvp: No such file or directory

while now it is:

    after: execvp(77790): No such file or directory

so it's more easy to guess the issue.

removed also dieerrno() because it wasn't suitable to print arguments of
a failed function.

bumped version.